feat(product): querydsl 적용 및 필요 없는 dto 삭제 및 정리
안쓰는 dto 삭제
Querydsl 설정
Querydsl 이용 쿼리 작성

+++ src/server/product/src/main/kotlin/com/cream/product/dto/PageDTO.kt
@@ -3,19 +3,19 @@ package com.cream.product.dto
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class PageDTO (
-    val cursor: Int,
+    val cursor: Long,
 
     @JsonProperty("per_page")
-    val perPage: Int,
+    val perPage: Long,
 
     val sort: String
     )
 {
-    fun offset(): Int{
+    fun offset(): Long{
         return cursor * perPage
     }
 
-    fun limit(): Int{
+    fun limit(): Long{
         return (cursor * perPage) + perPage
     }
 }
